add changes per request, devided in two components
