+++ src/components/pages/info-pages/ClassInfoPage.jsx
@@ -1,78 +1,14 @@
 import React, { useEffect, useState } from 'react';
-import { getClassDetails, getUserData } from '../../util/DataBaseRequests';
-import { useAuth } from '../../AuthProvider';
+import { getClassDetails, getUserData } from '../../../util/DataBaseRequests';
+import { useAuth } from '../../../AuthProvider';
 import { useParams } from 'react-router-dom';
-import Loader from '../common/Loader';
-import AgeIcon from '../../assets/icons/icon-age.svg';
-import LessonTypeIcon from '../../assets/icons/icon-lesson.svg';
-import ScheduleIcon from '../../assets/icons/icons-schedule.svg';
-import IconClock from '../../assets/icons/icon-clock.png';
-import IconDegree from '../../assets/icons/icons-degree.png';
-
-const TeacherInfo = ({ teacherInfo }) => {
-  const [showFullExperience, setShowFullExperience] = useState(false);
-  //handle ReadMore button
-  const handleToggleExperience = () => {
-    setShowFullExperience(!showFullExperience);
-  };
-  return (
-    <div className="mt-6">
-      <p className="text-3xl font-spartan font-medium  mb-4">
-        Meet the Teacher
-      </p>
-      <div className="flex flex-row">
-        {teacherInfo.profileImageUrl && (
-          <img
-            src={teacherInfo.profileImageUrl}
-            alt={`${teacherInfo.firstName} ${teacherInfo.lastName}`}
-            className="w-20 h-20 rounded-full"
-          />
-        )}
-        <div className="flex text-center items-center justify-center  pl-4">
-          <p className="font-roboto font-medium text-xl">{`${teacherInfo.firstName || ''} ${teacherInfo.lastName || ''}`}</p>
-        </div>
-      </div>
-      <div className="flex flex-row mt-4 gap-4">
-        {teacherInfo.profileImageUrl && (
-          <button className="bg-pureWhite py-1 px-4 hover:bg-red hover:text-pureWhite hover:border-2 hover:border-red text-red font-spartan font-semibold text-lg rounded-md border-2 border-red">
-            Profile
-          </button>
-        )}
-        <button className="bg-pureWhite py-1 px-4 hover:bg-red hover:text-pureWhite hover:border-2 hover-border-red text-red font-spartan font-semibold text-lg rounded-md border-2 border-red">
-          Send Message
-        </button>
-      </div>
-      <div className="mt-6">
-        {teacherInfo.education && (
-          <>
-            <p className="text-3xl font-spartan font-medium lg:my-4">
-              Teacher education and experience
-            </p>
-            <div className="flex flex-row gap-2">
-              <img src={IconDegree} alt="Icon degree" className="w-6 h-6" />
-              <p className="font-roboto text-xl">{teacherInfo.education}</p>
-            </div>
-          </>
-        )}
-        {teacherInfo.experience && (
-          <>
-            <p className="font-roboto leading-7 text-xl pt-4 w-2/3">
-              {showFullExperience
-                ? teacherInfo.experience
-                : `${teacherInfo.experience.substring(0, 100)}...`}
-            </p>
-            <button
-              onClick={handleToggleExperience}
-              className="text-black font-semibold underline hover:text-darkGreen"
-            >
-              {showFullExperience ? 'Show Less' : 'Read More'}
-            </button>
-          </>
-        )}
-      </div>
-    </div>
-  );
-};
+import Loader from '../../common/Loader';
+import AgeIcon from '../../../assets/icons/icon-age.svg';
+import LessonTypeIcon from '../../../assets/icons/icon-lesson.svg';
+import ScheduleIcon from '../../../assets/icons/icons-schedule.svg';
+import IconClock from '../../../assets/icons/icon-clock.svg';
+import IconTypeLesson from '../../../assets/icons/icon-type.png';
+import TeacherInfo from './TeacherInfo';
 
 const ClassInfoPage = () => {
   const { userData } = useAuth();
@@ -107,7 +43,7 @@ const ClassInfoPage = () => {
 
       fetchClassAndTeacher();
     } else {
-      console.error('No token found, user might not be logged in.');
+      error('User should log in.');
     }
   }, [classId, userData]);
 
@@ -116,7 +52,7 @@ const ClassInfoPage = () => {
   if (!classItem) return <p>Class not found.</p>;
 
   return (
-    <div className="container m-auto p-10">
+    <div className="container m-auto w-10/12">
       <div className="flex flex-col w-full text-black">
         <div className="flex justify-between flex-col lg:flex-row">
           <div className="flex flex-col w-full lg:w-[75%] md:flex-row lg:flex-row pb-4">
@@ -174,6 +110,15 @@ const ClassInfoPage = () => {
                   />
                   <span>On Request</span>
                 </div>
+
+                <div className="flex items-center p-2 gap-2">
+                  <img
+                    src={IconTypeLesson}
+                    alt="Icon age"
+                    className="w-6 h-6"
+                  />
+                  <span>{classItem.lessonType}</span>
+                </div>
               </div>
 
               <div className="flex justify-center mt-4">
@@ -218,7 +163,6 @@ const ClassInfoPage = () => {
           <p>No available times listed.</p>
         )}
       </div>
-
       {teacherInfo && <TeacherInfo teacherInfo={teacherInfo} />}
     </div>
   );
